requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bokeh==3.3.2
cachelib==0.12.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.6.3
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
contourpy==1.0.7
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.8
Cython==0.29.36
dataclasses-json==0.6.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==4.2
dnspython==2.6.1
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
eventlet==0.36.1
executing==1.2.0
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.0
Flask-Limiter==3.7.0
Flask-Session==0.6.0
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.6.1
gast==0.5.4
gcsfs==2023.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.4
ijson==3.2.3
imageio==2.34.0
importlib_resources==6.4.0
inference-sdk==0.16.3
ipykernel==6.20.2
ipython==8.9.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwt==1.3.1
kaggle==1.6.5
keras==2.14.0
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
libclang==18.1.1
limits==3.13.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.2.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip3-autoremove==1.2.2
pipreqs==0.4.13
platformdirs==2.6.2
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.14.0
pymongo==4.6.3
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-slugify==8.0.3
python-socketio==5.11.3
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.0.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.16.0
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
streamlit==1.38.0
supervision==0.22.0
sympy==1.12
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
texttable==1.7.0
tf_keras==2.16.0
thinc==8.2.5
threadpoolctl==3.2.0
tifffile==2024.2.12
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
torchsummary==1.5.1
torchtext==0.17.2
torchvision==0.17.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.44.2
trio==0.23.1
trio-websocket==0.11.1
tweepy==4.14.0
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.16
vaderSentiment==3.3.2
virtualenv==20.23.1
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.2.0
xyzservices==2023.10.1
yarg==0.1.9
yarl==1.9.2
